LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_capture_mouse-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Events capture test

Test Description: This test checks if setCapture/releaseCapture functions works properly. Complete the following actions:
Move your mouse over the black rectangle. pointermove event should be logged in the black rectangle
Move your mouse over the purple rectangle. pointerover event should be logged in the purple rectangle
Press and hold left mouse button over "Set Capture" button. "gotpointercapture" should be logged in the black rectangle
Move your mouse anywhere. pointermove should be logged in the black rectangle
Move your mouse over the purple rectangle. Nothig should happen
Move your mouse over the black rectangle. pointermove should be logged in the black rectangle
Release left mouse button. "lostpointercapture" should be logged in the black rectangle

Test passes if the proper behaviour of the events is observed.



Pointer Events Capture Test

The following pointer types were detected: (none).

The following events were logged: (none).


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN gotpointercapture event received
NOTRUN lostpointercapture event received
NOTRUN pointerover event for black rectangle received
NOTRUN pointerover event for purple rectangle received
